Works around a spill bug in the RA
Spilling is a mess in this version of the RA and there was a significant
bug that effectively made spilling either infinite loop or crash.

This adds a bunch of new comments about the logic of the spilling and
ipmroves it so it shouldn't break as easily.

There are still some edge cases that can be hit that break spilling but
a lot of this code is going to be scrapped for the new RA coming up
anyway.
This is temporary until the new RA merges in.