Enables C++20 feature level
There are a few nice features in C++20 that can make our life easier.
This isn't requiring the full C++20 suite of features, but instead
should allow us to start bringing in features that were at least in
Clang 10.

LLVM 11 came out yesterday (2020-10-12) so it would be a shame to have
that as a minimum requirement.