Changes FEXInterpreter install to a hardlink
The only way to determine if we are running in the "interpreter" mode is
by determining filename.
if any application does an execve then we use /proc/self/exe to
determine which FEX we should execute. Sadly this resolved the softlink
to FEXLoader, breaking the chain.
Change it to a hardlink in order to ensure this doesn't happen.

This also moves the install target outside of aarch64 install target so
it can be tested on x86 hosts.